+++ packages/create-openfort/src/cli/index.ts
@@ -4,5 +4,6 @@ export * from './utils'
 export * from './FileManager'
 export * from './colors'
 export * from './template'
+export * from './errors'
 
 export { prompts };